Genres component broken into smaller components
