+++ src/components/Navigation/MenuToggle.js
@@ -37,11 +37,3 @@ const Path = (props) => (
   />
 );
 
-// Finished code
-// export const MenuToggle = ({ toggle }) => (
-//   <button onClick={toggle}>
-//     <svg width="23" height="23" viewBox="0 0 23 23">
-
-//     </svg>
-//   </button>
-// );
