--- src/pages/Genres/board.js
@@ -1,11 +1,0 @@
-import React, { useState } from "react";
-import GenreCard from "../../components/GenreCard";
-
-export default function MyDraggingBoard() {
- 
-
-  return (
-    
-  );
-}
-
